packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-qa-runs"
version = "0.1.0"
description = "Test execution records, assertion results, and pass/fail tracking across clawspaces"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["qa-runs","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-qa-runs"
Repository = "https://github.com/mxn2020/minions-qa-runs"

[tool.hatch.build.targets.wheel]
packages = ["minions_qa_runs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
